fixed errors on linux

+++ src/core/hardware/label_map.hpp
@@ -2,9 +2,7 @@
 
 #include <unordered_map>
 
-#include "app/globals.hpp"
 #include "proto/hardware.pb.h"
-#include "utils/types.hpp"
 
 class LabelMap {
     std::unordered_map<std::string, ushort> label_map;
@@ -20,9 +18,7 @@ class LabelMap {
     std::string const& at(ushort address) const;
     ushort at(std::string const& label) const;
     bool contains(std::string const& label) const;
-    inline ulong size() const {
-        return static_cast<ulong>(label_map.size());
-    }
+    inline size_t size() const { return label_map.size(); }
     void clear();
     void get_addresses(std::vector<ushort>& out) const;
     google::protobuf::RepeatedPtrField<ant_proto::LabelRecord> get_proto()
