+++ tests/unit_tests/core_unit_test/hardware_unit_test/parser/test_parser.cpp
@@ -7,7 +7,7 @@
 
 struct ParseCase {
     std::vector<std::string> lines;
-    ulong expected_code_size;
+    size_t expected_code_size;
 };
 
 class ParserParseTest : public ::testing::TestWithParam<ParseCase> {};
